requirements.txt
fastapi
pdf2image
google-generativeai
uvicorn
fastapi